+++ resources/views/layouts/app.blade.php
@@ -51,6 +51,11 @@
                                     <a class="nav-link" href="{{ route('billing-periods.index') }}">Abrechnung</a>
                                 </li>
                             @endcan
+                            @can('viewAny', App\Models\BillingRateTemplate::class)
+                                <li class="nav-item">
+                                    <a class="nav-link" href="{{ route('billing-rate-templates.index') }}">Preisvorlagen</a>
+                                </li>
+                            @endcan
                             @can('viewAny', App\Models\Invoice::class)
                                 <li class="nav-item">
                                     <a class="nav-link" href="{{ route('invoices.index') }}">Rechnungen</a>